vim-patch:8.1.1384: using "int" for alloc() often results in compiler warnings
Problem:    Using "int" for alloc() often results in compiler warnings.
Solution:   Use "size_t" and remove type casts.  Remove alloc_check(), Vim
            only works with 32 bit ints anyway.
https://github.com/vim/vim/commit/964b3746b9c81e65887e2ac9a335f181db2bb592

N/A commits:

vim-patch:8.1.0228: dropping files is ignored while Vim is busy

Problem:    Dropping files is ignored while Vim is busy.
Solution:   Postpone the effect of dropping files until it's safe.
https://github.com/vim/vim/commit/92d147be959e689f8f58fd5d138a31835e160289

vim-patch:8.2.3040: GUI: dropping files not tested

Problem:    GUI: dropping files not tested.
Solution:   Add test_gui_drop_files() and tests. (Yegappan Lakshmanan,
            closes vim/vim#8434)
https://github.com/vim/vim/commit/18d46587b985923ef4b90b19a0cf37a094607fec

+++ src/nvim/change.c
@@ -625,7 +625,7 @@ void ins_char_bytes(char_u *buf, size_t charlen)
     }
   }
 
-  char_u *newp = xmalloc((size_t)(linelen + newlen - oldlen));
+  char_u *newp = xmalloc(linelen + newlen - oldlen);
 
   // Copy bytes before the cursor.
   if (col > 0) {
